requirements.txt
Pillow>=10.0.0
openpyxl>=3.1.0
numpy>=1.24.0
